pretty straightforward.
P2 is very slow because 2D array are not supported. It would be faster
if I keep updating an array with the size that I am looking for before
actually reading value from str element I guess. Plus, store the max
value of empty space might avoid a lot of iterations.